# openapi.yaml
openapi: 3.0.3
info:
  title: Project Overwatch API
  description: |
    GDG Hackathon Project: AI-Agent Cyber Exposure & Intelligence Platform for Rwanda.
    Acts as an early-warning radar for FinTech threats (e.g., Mobile Money/MoMo smishing).
    Zero-Trust pipeline ensures all IPv4 addresses are masked before reaching frontend or AI payloads.
  version: 1.0.0
  contact:
    name: Fidele
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/auth/admin-login:
    post:
      tags: [Auth]
      summary: Mock admin login
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/public/scan-link:
    post:
      tags: [Public]
      summary: Scan a URL to determine threat level (MoMo phishing)
      operationId: scanLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  example: http://momo-bonus-win.rw/login
      responses:
        '200':
          description: Scan results and recommendations
...
  /api/ai/remediation:
    post:
      tags: [AI]
      summary: Get AI remediation steps for a vulnerability
      operationId: getRemediation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemediationRequest'
      responses:
        '200':
          description: Remediation steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemediationResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Some error message
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: overwatch2026
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Admin login successful
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                  example: admin-001
                username:
                  type: string
                  example: admin
                role:
                  type: string
                  example: admin
            token:
              type: string
              example: mock-jwt-token-overwatch-2026
...
    VulnerabilityInput:
      type: object
      required: [city, lat, lng, riskScore, threatType, ipAddress, severity]
      properties:
        city:
          type: string
          enum: [Kigali, Musanze, Rubavu]
          example: Kigali
        lat:
          type: number
          example: -1.9441
        lng:
          type: number
          example: 30.0619
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
          example: 82
        threatType:
          type: string
          example: Phishing Domain
        ipAddress:
          type: string
          format: ipv4
          example: 197.243.10.45
        severity:
          type: string
          enum: [high, medium, low]
          example: high
...
    RemediationRequest:
      type: object
      required: [vulnerability, language]
      properties:
        vulnerability:
          $ref: '#/components/schemas/VulnerabilityInput'
        language:
          type: string
          enum: [en, rw]
          example: en
    RemediationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            prompt:
              type: string
              description: The AI prompt (IPs masked)
            language:
              type: string
              enum: [en, rw]
            threatType:
              type: string
            remediationSteps:
              type: array
              items:
                type: string
              minItems: 3
              maxItems: 3